new yum option: don't expand removals
Signed-off-by: William Herrin <herrin-chef@dirtside.com>

This patch creates a configuration option which can be placed in client.rb.
If set to true, chef will invoke yum with remove_leaf_only=1 and then
check if yum actually removed the packages. If yum did not, the chef run
fails.

By default, "yum -y remove" will delete as many packages as needed to
remove the one package it was told it to remove. When used interactively,
this is reasonable: present the user with the list to be removed and
ask if that's what they meant to do.

When used under chef, deleting more than the requested package is usually
a mistake: the recipe author didn't know that something they still want
requires the package they thought they didn't want. Ideally, yum would
fail and stop here rather than wreck the server by removing unintended
packages.

Yum has an option to prevent this behavior: "remove_leaf_only=1". It
doesn't behave as Chef expects. Yum refuses to remove the package with
the dependency but then reports status 0 back to chef - I'm happy, I did
as you asked even though I, well, didn't. Chef then tries (and silently
fails) to remove the package on every subsequent run.

Yum development is dead (replaced by dnf) so if we want Yum to do the right
thing for Chef we have to fix it in Chef.